moved collisions from DynamicCube to subclass CollidableDynamicCube
