+++ core/src/com/geometric/wars/player/PlayersCube.java
@@ -2,13 +2,12 @@ package com.geometric.wars.player;
 
 import com.badlogic.gdx.graphics.g3d.Environment;
 import com.badlogic.gdx.graphics.g3d.ModelBatch;
-import com.geometric.wars.collisions.DynamicBody;
-import com.geometric.wars.cube.DynamicCube;
+import com.geometric.wars.cube.CollidableDynamicCube;
 import com.geometric.wars.cube.DynamicCubeController;
 
-public abstract class PlayersCube implements DynamicBody {
+public abstract class PlayersCube {
     protected DynamicCubeController dynamicCubeController;
-    protected DynamicCube dynamicCube;
+    protected CollidableDynamicCube dynamicCube;
 
 
     public void setPosition(int x,int y) {
@@ -34,14 +33,4 @@ public abstract class PlayersCube implements DynamicBody {
         batch.render(dynamicCube.getView(), environment);
     }
 
-
-    @Override
-    public boolean canCollideWith(DynamicBody object) {
-        if(object instanceof PlayersController)
-            return false;
-        throw new RuntimeException("dynamicGameObject "+object+" not supported by PlayersController collisions");
-    }
-
-    @Override
-    public void onCollisionWith(DynamicBody object) {}
 }
